backend/requirements.txt
# Web framework
fastapi[standard]>=0.128.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.0

# RAG — document parsing (lightweight, no PyTorch)
unstructured>=0.16.0
python-docx>=1.0.0
python-pptx>=1.0.0
pypdf>=5.0.0

# Embeddings
tiktoken>=0.8.0

# Utilities
python-dotenv>=1.0.0

# Config
pydantic-settings>=2.0.0

# Agent core (Phase 4)
pydantic-ai[anthropic,openai,duckduckgo]>=1.0.0

# LLM SDKs
anthropic>=0.40.0
openai>=1.50.0

# SSE streaming
sse-starlette>=2.0.0

# Testing
pytest>=8.0.0
anyio[trio]>=4.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
aiosqlite>=0.20.0
